+++ dtslint/ts4.1/Either.ts
@@ -38,11 +38,11 @@ flow(
 )('foo')
 
 //
-// Witherable overlodings
+// Filterable overlodings
 //
 
 declare function isString(x: unknown): x is string
-const W = _.getWitherable(monoidAll)
+const W = _.getFilterable(monoidAll)
 
 W.filter(_.right<boolean, string | number>(1), isString) // $ExpectType Either<boolean, string>
 W.partition(_.right<boolean, string | number>(1), isString) // $ExpectType Separated<Either<boolean, string | number>, Either<boolean, string>>